# PERF: Improve replace perf