@jreback 2016-03-30
notice that ``BoolBlock.replace`` doesn't have ``convert`` arg at all! so not tested / working. can you give a check.