@sinhrks 2016-03-30
Ah, found some ``.replace`` conversion looks buggy. I'll submit a separate issue as the fix not looks straightforward.