@sinhrks 2016-03-30
See #12747. Change this PR title to WIP until #12747 is fixed.