@jreback 2016-11-23
lgtm. can we merge this independent of the other PR's? (so it could go in 0.19.2)