@sinhrks 2016-11-29
It is independent. #12747 fixes some existing bugs separately.